move cmd/zar to ppl/cmd/zar
