+++ README.md
@@ -8,7 +8,7 @@ and store structured log data, including:
 * The [zqd](ppl/cmd/zqd/README.md) daemon, which serves a REST-like API to manage
  and query log archives, and is the backend for the [Brim](https://github.com/brimsec/brim)
  application
-* The [zar](cmd/zar/README.md) command line tool, for working with log data
+* The [zar](ppl/cmd/zar/README.md) command line tool, for working with log data
  archives
 * The [zapi](cmd/zapi/README.md) command line tool, for interacting with the
 API provided by zqd
