+++ cmd/microindex/README.md
@@ -1,7 +1,7 @@
 # `microindex`
 
 microindex is a command-line utility for debugging and interrogating microindex ZNG files
-that are used as internal indexes for [zar](../zar/README.md) archives.
+that are used as internal indexes for [zar](../../ppl/cmd/zar/README.md) archives.
 
 For all `microindex` options, use the help subcommand:
