+++ ppl/cmd/zar/rmdirs/command.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/brimsec/zq/archive"
-	"github.com/brimsec/zq/cmd/zar/root"
+	"github.com/brimsec/zq/ppl/cmd/zar/root"
 	"github.com/mccanne/charm"
 )
