+++ ppl/cmd/zar/root/command.go
