+++ ppl/cmd/zar/stat/command.go
@@ -7,8 +7,8 @@ import (
 	"os"
 
 	"github.com/brimsec/zq/archive"
-	"github.com/brimsec/zq/cmd/zar/root"
 	"github.com/brimsec/zq/emitter"
+	"github.com/brimsec/zq/ppl/cmd/zar/root"
 	"github.com/brimsec/zq/zbuf"
 	"github.com/brimsec/zq/zio"
 	"github.com/brimsec/zq/zng/resolver"
